requirements.txt
google-api-python-client==2.184.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
# Transitive dependencies pinned for reproducibility (optional)
google-auth==2.41.1
httplib2==0.31.0
uritemplate==4.2.0